serviceability: exchange.bgp_community u32 -> u16

+++ smartcontract/sdk/rs/src/commands/exchange/update.rs
@@ -12,7 +12,7 @@ pub struct UpdateExchangeCommand {
     pub name: Option<String>,
     pub lat: Option<f64>,
     pub lng: Option<f64>,
-    pub bgp_community: Option<u32>,
+    pub bgp_community: Option<u16>,
 }
 
 impl UpdateExchangeCommand {
